refactor: CSR 로직 컴포넌트로 분리 및 ESLint 오류 참고해 코드 수정
