+++ src/components/ArticleList.module.css
@@ -1,0 +1,47 @@
+.post_action {
+  display: flex;
+  justify-content: space-between;
+  padding: 24px 0;
+}
+
+.post_header button:hover {
+  background-color: #1967D6;
+}
+
+.search_container {
+  position: relative;
+  width: 1054px;
+  height: 42px;
+}
+
+.searchIcon {
+  position: absolute;
+  left: 16px;
+  top: 50%;
+  transform: translateY(-50%);
+}
+
+.search_bar {
+  width: 100%;
+  height: 100%;
+  padding: 9px 20px 9px 40px;
+  border: none;
+  border-radius: 12px;
+  background-color: #f3f4f6;
+  font-size: 16px;
+  line-height: 26px;
+  color: #9ca3af;
+  outline: none;
+}
+
+.all_section {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+@media (max-width: 1199px) {
+  .post_action {
+    gap: 15px;
+  }
+}
