+++ src/components/BoardHeader.module.css
@@ -1,0 +1,24 @@
+.post_header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding-top: 40px;
+}
+
+.post_header h3 {
+  font-size: 20px;
+  line-height: 32px;
+  font-weight: 700;
+  color: #1f2937;
+}
+
+.post_header button {
+  padding: 12px 23px;
+  border: none;
+  border-radius: 8px;
+  background-color: #3692ff;
+  color: #ffffff;
+  font-weight: 600;
+  line-height: 19px;
+  cursor: pointer;
+}
