+++ src/pages/boards/index.module.css
@@ -17,78 +17,7 @@
   max-width: 1fr;
 }
 
-.post_header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-top: 40px;
-}
-
-.post_header h3 {
-  font-size: 20px;
-  line-height: 32px;
-  font-weight: 700;
-  color: #1f2937;
-}
-
-.post_header button {
-  padding: 12px 23px;
-  border: none;
-  border-radius: 8px;
-  background-color: #3692ff;
-  color: #ffffff;
-  font-weight: 600;
-  line-height: 19px;
-  cursor: pointer;
-}
-
-.post_action {
-  display: flex;
-  justify-content: space-between;
-  padding: 24px 0;
-}
-
-.post_header button:hover {
-  background-color: #1967D6;
-}
-
-.search_container {
-  position: relative;
-  width: 1054px;
-  height: 42px;
-}
-
-.searchIcon {
-  position: absolute;
-  left: 16px;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.search_bar {
-  width: 100%;
-  height: 100%;
-  padding: 9px 20px 9px 40px;
-  border: none;
-  border-radius: 12px;
-  background-color: #f3f4f6;
-  font-size: 16px;
-  line-height: 26px;
-  color: #9ca3af;
-  outline: none;
-}
-
-.all_section {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
 @media (max-width: 1199px) {
-  .post_action {
-    gap: 15px;
-  }
-
   .best_section> :nth-child(3) {
     display: none;
     width: 100%;
